refactor: split module file
